requirements.txt
Flask>=1.1.1
flask-restplus>=0.13.0
Flask-SQLAlchemy>=2.4.1
Flask-JWT-Extended>=3.23.0
itsdangerous
Flask-Mail
flask-migrate>=2.5.2
Flask-Script>=2.0.6
flask-cors
Flask-Testing
psycopg2-binary
requests
flask-marshmallow
marshmallow-sqlalchemy
marshmallow_enum
nose
gunicorn
locustio
